Update fromcd.sh
Fixed an error where files were renamed before metadata was collected. So the file would not be found.

Increased the matching requirements on renaming chapter titles. 
Now more specific search for Chapter at the beginning of a track
Will also match just a space after the chapter number and change to ". " (period space)